requirements.txt
# Mutated Protein Modeling Pipeline - Dependencies

# Core scientific libraries
numpy>=1.21.0
scipy>=1.7.0

# Bioinformatics
biopython>=1.79

# Machine Learning / PCA
scikit-learn>=1.0.0

# HTTP requests
requests>=2.25.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
py3Dmol>=1.8.0

# BioEmu (conformational sampling)
# Note: BioEmu requires GPU (CUDA)
bioemu

# PyTorch (required by BioEmu)
# Install appropriate version for your CUDA setup:
# torch>=1.10.0
# See: https://pytorch.org/get-started/locally/

# Optional: Jupyter support (for notebook version)
# jupyter>=1.0.0
# ipywidgets>=7.6.0
